refactor : TimeStamp를 공통으로 사용하기 위해 구조 변경
